Fixed circular dependency issues
